try to organize nmod_mpoly_factor/gcd_zippel a bit
