+++ n_poly.h
@@ -191,6 +191,8 @@ void n_poly_clear(n_poly_t A)
         flint_free(A->coeffs);
 }
 
+FLINT_DLL int n_poly_is_canonical(const n_poly_t A);
+
 FLINT_DLL void n_poly_realloc(n_poly_t A, slong len);
 
 FLINT_DLL void n_poly_print_pretty(const n_poly_t A, const char * x);
@@ -1819,6 +1821,8 @@ void n_polyun_init(n_polyun_t A)
     A->length = 0;
     A->alloc = 0;
 }
+
+FLINT_DLL int n_polyun_is_canonical(const n_polyun_t A);
 
 FLINT_DLL void n_polyun_clear(n_polyun_t A);
