+++ n_poly/n_polyu.c
@@ -9,10 +9,28 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
-#include "nmod_mpoly_factor.h"
+#include "n_poly.h"
 #include "mpn_extras.h"
 #include "nmod_vec.h"
+#include "mpoly.h"
 
+int n_polyun_is_canonical(const n_polyun_t A)
+{
+    slong i;
+    if (A->length < 0)
+        return 0;
+    for (i = 0; i < A->length; i++)
+    {
+        if (!n_poly_is_canonical(A->terms[i].coeff) ||
+            n_poly_is_zero(A->terms[i].coeff))
+        {
+            return 0;
+        }
+        if (i > 0 && A->terms[i].exp >= A->terms[i - 1].exp)
+            return 0;
+    }
+    return 1;
+}
 
 void n_polyu_clear(n_polyu_t A)
 {
